requirements.txt
autopep8==1.5.2
beautifulsoup4==4.9.0
bs4==0.0.1
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.1
dnspython==1.16.0
Flask==1.1.2
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mysql-connector-python==8.0.19
protobuf==3.6.1
pycodestyle==2.5.0
python-dotenv==0.13.0
requests==2.23.0
six==1.14.0
soupsieve==2.0
urllib3==1.25.9
Werkzeug==1.0.1
